pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["ribeyra <ribeyra@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
python-dotenv = "^1.0.1"
dj-database-url = "^2.2.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^22.0.0"
uvicorn = "^0.30.1"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
django-bootstrap5 = "^24.2"
django-filter = "^24.2"


[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.3"
ipython = "^8.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
